Re-organized components and added styling to labels
