+++ src/components/MainForm/FormInput/FormInput.css
@@ -1,0 +1,5 @@
+.labelStyle {
+  font-family: -apple-system,BlinkMacSystemFont,"Segoe UI",Roboto,"Helvetica Neue",Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol";
+  line-height: 1.5;
+  font-weight: 500;
+}
