+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import MainForm from './MainForm/MainForm';
+import MainForm from './components/MainForm/MainForm';
 import registerServiceWorker from './registerServiceWorker';
 
 ReactDOM.render(<MainForm />, document.getElementById('root'));
